+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 ---------
 <!--(CHANGELOG_TOP)-->
+**1.2.2**
+* added error events to be sent for invalid custom event fields used
+* added optional mergeFields argument to event methods to merge with global custom fields instead of overwrite them
+
 **1.2.1**
 * fixed missing custom event fields for when trying to fix missing session end events